Move wav2vec2 ASR metrics to their own module
