+++ src/fairseq2/models/wav2vec2/asr/__init__.py
@@ -15,14 +15,8 @@ from fairseq2.models.wav2vec2.asr.factory import wav2vec2_asr_arch as wav2vec2_a
 from fairseq2.models.wav2vec2.asr.factory import (
     wav2vec2_asr_archs as wav2vec2_asr_archs,
 )
-from fairseq2.models.wav2vec2.asr.model import (
-    Wav2Vec2AsrMetricBag as Wav2Vec2AsrMetricBag,
-)
 from fairseq2.models.wav2vec2.asr.model import Wav2Vec2AsrModel as Wav2Vec2AsrModel
 from fairseq2.models.wav2vec2.asr.model import Wav2Vec2AsrOutput as Wav2Vec2AsrOutput
-from fairseq2.models.wav2vec2.asr.model import (
-    Wav2Vec2AsrValidMetricBag as Wav2Vec2AsrValidMetricBag,
-)
 from fairseq2.models.wav2vec2.asr.setup import (
     load_wav2vec2_asr_config as load_wav2vec2_asr_config,
 )
